metal: Move "computed_macros" to "ember-runtime"
