+++ packages/ember-routing-views/lib/components/link-to.js
@@ -316,7 +316,7 @@ import Logger from 'ember-metal/logger';
 import { assert, deprecate } from 'ember-metal/debug';
 import { get } from 'ember-metal/property_get';
 import { computed } from 'ember-metal/computed';
-import { deprecatingAlias } from 'ember-metal/computed_macros';
+import { deprecatingAlias } from 'ember-runtime/computed/computed_macros';
 import { isSimpleClick } from 'ember-views/system/utils';
 import EmberComponent from 'ember-views/components/component';
 import inject from 'ember-runtime/inject';
